Moved UserPayload type to another directory
